refactor to make model folder structure consistent
