+++ torchtitan/experiments/flux/__init__.py
@@ -9,13 +9,14 @@
 
 from torchtitan.components.lr_scheduler import build_lr_schedulers
 from torchtitan.components.optimizer import build_optimizers
-from torchtitan.experiments.flux.dataset.flux_dataset import build_flux_dataloader
-from torchtitan.experiments.flux.loss import build_mse_loss
-from torchtitan.experiments.flux.model.autoencoder import AutoEncoderParams
-from torchtitan.experiments.flux.parallelize_flux import parallelize_flux
 from torchtitan.protocols.train_spec import register_train_spec, TrainSpec
 
-from .model.model import FluxModel, FluxModelArgs
+from .dataset.flux_dataset import build_flux_dataloader
+from .infra.parallelize import parallelize_flux
+from .loss import build_mse_loss
+from .model.args import FluxModelArgs
+from .model.autoencoder import AutoEncoderParams
+from .model.model import FluxModel
 
 __all__ = [
     "FluxModelArgs",
