+++ torchtitan/models/llama3/model/model.py
@@ -7,79 +7,14 @@
 # Copyright (c) Meta Platforms, Inc. All Rights Reserved.
 
 
-from dataclasses import dataclass
-
 import torch
 import torch.nn.functional as F
 from torch import nn
 
-from torchtitan.components.tokenizer import Tokenizer
-from torchtitan.config_manager import JobConfig
 from torchtitan.models.attention import build_attention, init_attention_mask
-from torchtitan.protocols.train_spec import BaseModelArgs, ModelProtocol
+from torchtitan.protocols.train_spec import ModelProtocol
 
-
-@dataclass
-class TransformerModelArgs(BaseModelArgs):
-    dim: int = 4096
-    n_layers: int = 32
-    n_heads: int = 32
-    n_kv_heads: int | None = None
-    vocab_size: int = -1  # defined later by tokenizer
-    multiple_of: int = 256  # make SwiGLU hidden layer size multiple of large power of 2
-    ffn_dim_multiplier: float | None = None
-    norm_eps: float = 1e-5
-    rope_theta: float = 10000
-
-    max_seq_len: int = 2048
-    # If `True`, then each transformer block init uses its layer ID, and if
-    # `False`, each uses the total number of transformer blocks
-    depth_init: bool = True
-
-    use_flex_attn: bool = False
-    attn_mask_type: str = "causal"
-    eos_id: int = 0
-
-    def update_from_config(self, job_config: JobConfig, tokenizer: Tokenizer) -> None:
-        self.vocab_size = tokenizer.n_words
-        self.max_seq_len = job_config.training.seq_len
-        self.eos_id = tokenizer.eos_id
-
-        if job_config.activation_checkpoint.mode == "selective" and self.use_flex_attn:
-            raise ValueError(
-                "FlexAttention is not compatible with selective AC yet. "
-                "See https://github.com/pytorch/pytorch/issues/147879"
-            )
-
-        if job_config.parallelism.context_parallel_degree > 1 and self.use_flex_attn:
-            raise ValueError(
-                "FlexAttention is not compatible with CP yet. "
-                "We are still working on this."
-            )
-
-    def get_nparams_and_flops(self, model: nn.Module, seq_len: int) -> tuple[int, int]:
-        nparams = sum(p.numel() for p in model.parameters())
-        nparams_embedding = sum(
-            sum(p.numel() for p in m.parameters())
-            for m in model.children()
-            if isinstance(m, nn.Embedding)
-        )
-
-        l, h, q, t = (
-            self.n_layers,
-            self.n_heads,
-            self.dim // self.n_heads,
-            seq_len,
-        )
-        # Reasoning behind the factor of 12 for the self-attention part of the formula:
-        # 1. each self-attention has 2 matmul in the forward and 4 in the backward (6)
-        # 2. the flash attention does 1 more matmul recomputation in the backward
-        #    but recomputation should not be counted in calculating MFU           (+0)
-        # 3. each matmul performs 1 multiplication and 1 addition                 (*2)
-        # 4. we follow the convention and do not account for sparsity in causal attention
-        num_flops_per_token = 6 * (nparams - nparams_embedding) + 12 * l * h * q * t
-
-        return nparams, num_flops_per_token
+from .args import TransformerModelArgs
 
 
 def precompute_freqs_cis(dim: int, end: int, theta: float = 10000.0) -> torch.Tensor:
@@ -495,17 +430,3 @@ class Transformer(nn.Module, ModelProtocol):
         h = self.norm(h) if self.norm else h
         output = self.output(h) if self.output else h
         return output
-
-    @classmethod
-    def from_model_args(cls, model_args: TransformerModelArgs) -> "Transformer":
-        """
-        Initialize a Transformer model from a TransformerModelArgs object.
-
-        Args:
-            model_args (TransformerModelArgs): Model configuration arguments.
-
-        Returns:
-            Transformer: Transformer model.
-
-        """
-        return cls(model_args)
